For revision - Try w/ different compiler
I have implemented the dynamic array, however, I get an internal linkage error when compiling. 
Since this is pretty much the same error I had back with submission 1 (and it turned out to be a problem of the compiler for WSL), please download the files and try to compile it with your machine.